Add KnockoutJS context to DevTools sidebar
This commit introduces a new functionality to the DevTools by creating a sidebar pane that displays the KnockoutJS context to the selected DOM element. It adds a devtools.js file responsible for this added functionality along with its associated files – devtools.html and manifest.json. A .gitignore file is also added to ignore unnecessary files.